pyproject.toml
[project]
name = "student-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.52.0",
    "streamlit>=1.51.0",
]
